Upravljanje artiklima - ispravak

+++ Projekt/Skladiste_ETI/Skladiste_ETI/stavke.cs
@@ -17,7 +17,7 @@ namespace Skladiste_ETI
         public int id_stavke { get; set; }
         public Nullable<int> dokument_id_dokumenta { get; set; }
         public Nullable<int> artikli_id_artikla { get; set; }
-        public float kolicina { get; set; }
+        public double kolicina { get; set; }
         public double masa { get; set; }
     
         public virtual artikli artikli { get; set; }
